refactor: split code between internal app, web and gateway packages
