+++ internal/web/container_image_to_html.go
@@ -17,7 +17,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ******************************************************************************/
 
-package html_rendering
+package web
 
 import "strings"
